When downloading a file, correctly resolve it's path.
When action=dl (i.e. a file is being requested for download), the call
to &fullPath received undef, not $path, which is resolved out to point
to the $store and $point only, meaning the file being downloaded was
actually Perl's representation of the directory on the disk (!).

This patch also adjusts the returned Content-Description to be _more_
descriptive.

Fixes #1.